Fix en las migraciones
Se realizan correcciones a las migraciones

+++ database/migrations/2018_12_19_022841_create_user_detail_table.php
@@ -17,7 +17,7 @@ class CreateUserDetailTable extends Migration
             $table->increments('id')
                 ->comment('Código del detalle del usuario');
 
-            $table->unsignedInteger('user_id')                
+            $table->unsignedBigInteger('user_id')                
                 ->default('1')
                 ->comment('Código de Usuario');
